requirements.txt
Django==5.0.6
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
anthropic==0.29.0
python-decouple==3.8
Pillow==10.4.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
whitenoise==6.7.0
django-widget-tweaks==1.5.0